feat(sublime): Add `focus active pane` feature to sublime.
This is a custom plugin, which currently applies `<theme name>.inactive.<theme ext>` if it exists,
to any inactive window. Eventually I'd like for this to use the `ThemeTweaker` plugin to apply
filters to inactive themes. This would simply things greatly. Only problem is I couldn't get
ThemeTweaker to work with *.sublime-color-scheme files, even though it should support it. I have
opened the issue below on github to address this, after which we can look at refactoring.

REF: facelessuser/ThemeTweaker#12